+++ src/main/java/mil/nga/geopackage/tiles/user/TileRow.java
@@ -34,6 +34,17 @@ public class TileRow extends UserRow<TileColumn, TileTable> {
 	}
 
 	/**
+	 * Copy Constructor
+	 * 
+	 * @param tileRow
+	 *            tile row to copy
+	 * @since 1.3.0
+	 */
+	public TileRow(TileRow tileRow) {
+		super(tileRow);
+	}
+
+	/**
 	 * Get the zoom level column index
 	 * 
 	 * @return zoom level column index