+++ CHANGELOG.md
@@ -4,9 +4,12 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 1.2.3 (TBD)
+## [1.3.0](https://github.com/ngageoint/geopackage-java/releases/tag/1.3.0) (06-27-2017)
 
-* TBD
+* geopackage-core version updated to 1.3.0
+* tiff version updated to 1.0.3
+* Copy constructors for user table (features, tiles, attributes) row objects
+* Improved date column support for user tables (features, tiles, attributes)
 
 ## [1.2.2](https://github.com/ngageoint/geopackage-java/releases/tag/1.2.2) (06-12-2017)